Orpheus-FastAPI/tara-orpheus-livekit/requirements.txt
# livekit~=1.0.0
# livekit-agents==1.0.2
# livekit-plugins-openai~=1.0.0
# # optional, only when using LiveKit's turn detection model
# livekit-plugins-turn-detector
# # optional, only if background voice & noise cancellation is needed
# livekit-plugins-noise-cancellation>=0.2.0,<1.0.0
# python-dotenv

# livekit~=1.0.0
# livekit-agents==1.0.2
# livekit-plugins-openai~=1.0.0
# # optional, only when using LiveKit's turn detection model
# livekit-agents[turn-detector]~=1.0
# # optional, only if background voice & noise cancellation is needed
# livekit-plugins-noise-cancellation>=0.2.0,<1.0.0
# python-dotenv




livekit-agents~=1.0
livekit-agents[openai]~=1.0
livekit-plugins-openai~=1.0
livekit-plugins-cartesia~=1.0
livekit-plugins-deepgram~=1.0
livekit-plugins-silero~=1.0
livekit-plugins-elevenlabs~=1.0
# optional, only when using LiveKit's turn detection model
livekit-plugins-turn-detector~=1.0
# optional, only if background voice & noise cancellation is needed
livekit-plugins-noise-cancellation>=0.2.0,<1.0.0
python-dotenv~=1.0
